# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic E-Commerce & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /erp/stores:
    get:
      operationId: getErpStores
      tags: [erp]
      summary: List all stores (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createErpStore
      tags: [erp]
      summary: Create a new store
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /erp/stores/all:
    get:
      operationId: getErpStoresAll
      tags: [erp]
      summary: List all active tenant stores (any staff)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
  /erp/stores/mine:
    get:
      operationId: getErpStoresMine
      tags: [erp]
      summary: Stores the current user belongs to (with profile fields)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
  /erp/stores/{id}:
    patch:
      operationId: updateErpStore
      tags: [erp]
      summary: Update a store (rename / activate / deactivate)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteErpStore
      tags: [erp]
      summary: Delete a store (only if it has no data)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Store:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        tvaRate:
          type: string
          nullable: true
        showTvaByDefault:
          type: boolean
          nullable: true
        nif:
          type: string
          nullable: true
        rc:
          type: string
          nullable: true
        ai:
          type: string
          nullable: true
        defaultComptoirCustomerId:
          type: integer
          nullable: true
      required: [id, nameAr, nameEn, slug]
...
    CreateStoreRequest:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        slug:   { type: string }
        isActive: { type: boolean }
        address: { type: string }
        phone: { type: string }
        logoUrl: { type: string }
        tvaRate: { type: string }
        showTvaByDefault: { type: boolean }
        nif: { type: string }
        rc: { type: string }
        ai: { type: string }
      required: [nameAr, nameEn, slug]
    UpdateStoreRequest:
      type: object
      properties:
        nameAr:   { type: string }
        nameEn:   { type: string }
        isActive: { type: boolean }
        address: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        tvaRate: { type: string }
        showTvaByDefault: { type: boolean }
        nif: { type: string, nullable: true }
        rc: { type: string, nullable: true }
        ai: { type: string, nullable: true }
        defaultComptoirCustomerId: { type: integer, nullable: true }
...